Allow the prefetch input to be set for each architecture
On the previous commit, the prefetch was being executed individually per
architecture, but the same content was being fetched for all of them.

This commit changes it so that meaningful content filtering is possible
in each of the prefetch task runs, making it so that only the
arch-specific content is available for each build task run.

Signed-off-by: Bruno Pimentel <bpimente@redhat.com>